+++ docs/conf.py
@@ -19,7 +19,7 @@ extensions = [
 ]
 intersphinx_mapping = {
     "python": ("https://docs.python.org/3/", None),
-    "flask": ("http://flask.pocoo.org/docs/", None),
+    "flask": ("https://flask.palletsprojects.com/", None),
     "sqlalchemy": ("https://docs.sqlalchemy.org/en/latest/", None),
 }
 issues_github_path = "pallets/flask-sqlalchemy"
@@ -29,13 +29,15 @@ issues_github_path = "pallets/flask-sqlalchemy"
 html_theme = "flask"
 html_context = {
     "project_links": [
-        ProjectLink("Donate to Pallets", "https://palletsprojects.com/donate"),
-        ProjectLink("Website", "https://palletsprojects.com/"),
-        ProjectLink("PyPI releases", "https://pypi.org/project/Flask-SQLAlchemy/"),
+        ProjectLink("Donate", "https://palletsprojects.com/donate"),
+        ProjectLink("PyPI Releases", "https://pypi.org/project/Flask-SQLAlchemy/"),
         ProjectLink("Source Code", "https://github.com/pallets/flask-sqlalchemy/"),
         ProjectLink(
             "Issue Tracker", "https://github.com/pallets/flask-sqlalchemy/issues/"
         ),
+        ProjectLink("Website", "https://palletsprojects.com/"),
+        ProjectLink("Twitter", "https://twitter.com/PalletsTeam"),
+        ProjectLink("Chat", "https://discord.gg/pallets"),
     ]
 }
 html_sidebars = {
@@ -52,5 +54,5 @@ html_show_sourcelink = False
 # LaTeX ----------------------------------------------------------------
 
 latex_documents = [
-    (master_doc, f"Flask-SQLAlchemy-{version}.tex", html_title, author, "manual",)
+    (master_doc, f"{project}-{version}.tex", html_title, author, "manual")
 ]